Add slight ("hidden") variation on day 9.
Following some of the clarification posts on /r/adventofcode:

https://www.reddit.com/r/adventofcode/comments/5hc60x/what_does_your_day9part2_code_output_for_this/
https://www.reddit.com/r/adventofcode/comments/5hc77i/day_9_need_clarification/

It appears that the intention was for the repeat to happen before the decompress step
in v2 (rather than decompressing, then repeating).

Doing this, makes my code seems to run for **A LONG TIME** on the real input compared
to my original decompress-then-repeat approach, which seems non-useful.

Test data seems to return the same no matter which way it occurs, however the invalid
test data in the above 2 posts ends up in an infinite decompress loop or a seg-fault
if we repeat-then-decompress, rather than working fine in the decompress-then-repeat
approach.

Given that the inputs for the challenges are specifically curated and the original
method seems to work for them all, I'm leaving that as the default.